ENG-14787: Extract StackTrace to its own file
When the logger was implemented as a pure macro a cycle was created
between debuglog.h and ThreadLocalPool.h. This was resolved by removing
the include of debuglog.h from ThreadLocalPool.h. However, the
StackTrace class is used by ThreadLocalPool.h when pool checking is
enabled.

In order to keep the logging as a macro and still use the StackTrace in
ThreadLocalPool.h, StackTrace is being extracted to its own header and
source file. This allows both debuglog.h and ThreadLocalPool.h to
include StackTrace.h and not have a cycle.
